requirements.txt
# Core dependencies
langchain==0.1.16
langchain-community==0.0.34
ollama==0.1.8
pydantic==2.7.0

# Game and calculation
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1

# Visualization and UI
streamlit==1.32.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1

# Configuration and logging
pyyaml==6.0
python-dotenv==1.0.0
loguru==0.7.0

# Development and testing
pytest==7.4.0
black==23.3.0
flake8==6.0.0
